sgw: add SmartGatewayMaxCostMaxEtx configuration setting
SmartGatewayMaxCostMaxEtx: When a node advertises the maximum bandwidth
and its ETX is below the value of this setting then the resulting gateway
costs are equal to the ETX, otherwise the normal calculation of the
gateway costs applies (default is 2560).

This avoids instability in the mesh when a far-away node advertises
maximum bandwidth.

Signed-off-by: Ferry Huberts <ferry.huberts@pelagic.nl>