nameservice: rename parameter "sighup_interval" to "filewrite_interval"
Playing around with the new option "sighup_interval" I found, that this
also controls the interval of calling the "name-change-script".
Checking the code points out that this is the maintimer for writing
any status-file to disk. So it's more consistent to change the parameter-
name to what it's really doing.

Also update the README to reference this parameter in the others
affected by it.

Signed-off-by: Sven Roederer <freifunk@it-solutions.geroedel.de>