hello reception: remove duplicate neighbors
See the explanations in the descriptions of commits 2f38613 and
5089a79.

To fix the issue more thoroughly we need to ensure that no duplicate
neighbours are sent or received because that confuses the receiver.
This is especially true when the HELLO messages are fragmented, and
even more so when the fragments have some delay between them since
that causes link instability.

This commit fixes that for the receiving side.

Signed-off-by: Iwan G. Flameling <iwanovich@gmail.com>
Signed-off-by: Ferry Huberts <ferry.huberts@pelagic.nl>