fix segfault when printing "You must be root"
When olsr_cnf is NULL, always print OLSR_PRINTF messages to avoid segfault.

I compiled olsrd and ran ./olsrd --help, resulting in a segfault.
That's because main.c prints an error message even before the config is loaded,
so that OLSR_PRINTF tries to lookup the logging level in a NULL olsr_cnf.

Signed-off-by: Neels Hofmeyr <neels@hofmeyr.de>