fix: don't attempt to cleanup uninitialized multiGwTunnels
If olsr_cnf->smart_gw_use_count is 1, no multi-gw tunnels are initialized, and
the sgwTunnel4InterfaceNames and sgwTunnel6InterfaceNames are actually NULL. Do
not attempt to cleanup those if NULL.

Symptom: stopping olsrd e.g. via SIGINT leads to a segfault during cleanup.
The reason is that smart_gw_use_count is 1, which attempts to then dereference
one of sgwTunnel{4,6}InterfaceNames[0], amounting to <NULL>[0].

Signed-off-by: Neels Hofmeyr <neels@hofmeyr.de>