Fix adding IPv6 routes on BSD.
The KAME hack (storing the interface index inside the prefix) is only used
for link-local addresses. So only apply it in case the gateway is link-local.

For gateway routes (reachable on-link), install a cloning route which tells
the kernel to perform NDP address resolution to find the gateway.

The previous code was installing a bogus route to fe:80::1 for any destination.
I don't see how this could ever have worked.

Tested in Berlin's Freifunk mesh.

Signed-off-by: Stefan Sperling <stsp@stsp.name>