Use correct RTM flags while deleting a route on BSD.
If a route to a given destination already exists, the add_del_route()
function attempts to delete this route before inserting a new one.
While doing so, it asks olsr_rt_flags() for route message flags to use.

It should ask for route message flags that apply to a deletion. However,
it only did so if the 'add' parameter is false. This looks like a case of
badly copy-pasted code, since this section of code is obviously trying
to delete a route regardless of the value of 'add'.

Signed-off-by: Stefan Sperling <stsp@stsp.name>