add SmartGatewayBlacklist: implement blacklisting
Lookup function and config parsing were added in previous commits, this
actually implements the decision to avoid a smart gateway based on the
configured blacklist entries:

Add function isGwBlacklisted() to return nonzero when a given gw address
matches one of the blacklist entries, and to output debug info on level 2 that
a gateway was not considered, and which blacklist entry was responsible.

In gw_default_choose_gateway(), use isGwBlacklisted() to avoid any smartgw
matching an entry in the blacklist.

Rationale: so far olsrd may be stuck on specific smart gws without any means to
avoid broken or overloaded ones; any such broken gw not delivering on its
uplink promises can attract numerous clients' routing (e.g. because it has less
hops), thus actively disrupting service across a mesh segment. This patch
allows avoiding such service disruption by excluding gateways that are known to
be broken without having to rely on unknown/unreachable/too busy maintainers of
that node to fix it.

Signed-off-by: Neels Hofmeyr <neels@hofmeyr.de>